// src/api/auth/auth.type.ts
export type EmailResending = {
    email: string
}

export type BodyPasswordRecoveryCode = {
	newPassword: string;
	recoveryCode: string;
  };

  export type bodyAuthModel = {
	loginOrEmail: string
	password: string
}

export type BodyRegistrationConfirmationModel = {
	code: string
}

export type BodyRegistrationModel = {
	login: string
    email: string
    password: string
}